fix: use markdown_v2 msgtype in _build_markdown_v2_payload
The _build_markdown_v2_payload() method was returning 'markdown' as the
msgtype instead of 'markdown_v2'. This caused the payload.update() in
assemble_data() to overwrite the correct msgtype value.

This commit changes the msgtype in the builder method to 'markdown_v2'
to ensure the correct message type is sent to the WeChat Work API.

Signed-off-by: zhuoxyang <894647985@qq.com>